refactor(gateway): extract free-chat turn into FreeChatTurnRunner (v0.13 tranche 3)
Adapter shrinks **1411 → 962 lines (-449)** — first dip below 1000.
The biggest single concern in the file (full turn orchestration:
seed + retrieval + prune + first-call + mra-ask round + escalate
+ reply) moves into a 509-line focused module. 349/349 tests
still pass; the v0.13 harness's DM happy-path (3) + mra-ask
escalate (3) + msg_too_long hardening (3) tests transitively
cover every branch.

### New module `packages/cli/src/gateway/slack/free-chat-turn.ts` (509 lines)

`FreeChatTurnRunner` owns the entire free-chat orchestration:

- **`run(args)`** (fka `runFreeChatTurn`) — single turn end-to-end:
  PKB seed (first turn) → retrieval atom injection → pre-call
  prune (v0.11.1) → `chatWithContextRetry` → optional mra-ask
  round → optional escalate hand-off → strip directives, persist,
  update Slack placeholder.
- **`handleMraAskRound(args)`** (private, fka same name) — mra
  doctor check → progress-throttled subprocess launch → mra-ask
  audit event → call into `synthesiseAfterMra`.
- **`synthesiseAfterMra(args)`** (private, fka same name) — push
  preamble + mra-result into session, `chatWithContextRetry`
  with `phase="synthesise"`, return synthesised reply +
  scissors-prefix.

The `FreeChatSession` interface moves alongside its only consumer
(re-exported via `import type { FreeChatSession }` from this
module). Constructor takes `{ web, config, onLog, llm, mraDoctor,
runMraAsk, escalation }` — the runner needs the v0.13 tranche-2
EscalationCoordinator so the `escalate` directive branch can hand
off without going back through the adapter.

### SlackAdapter changes

- Drops 3 methods + their JSDoc (~420 lines).
- Drops local `FreeChatSession` interface (moved).
- Drops 15+ imports now exclusive to free-chat-turn:
  `chatWithContextRetry`, `PmkContextTooLongError`,
  `pickAudience`, `pickGatewayPrompt`, `createLastLineThrottle`,
  `sanitizeProgressLine`, `parseMraAsk`, `stripMraAskBlock`,
  `parseEscalate`, `stripEscalateBlock`, `searchAtoms`,
  `formatAtomsForInjection`, `buildIngestSeed`,
  `buildMraFailureMessage`, `buildMraSuccessMessage`,
  `capMessageContent`, `MRA_RESULT_CAP`, `pruneSessionIfNeeded`,
  `SEED_CAP`, `truncate`.
- Keeps `approxTokensFor` (still used by `handleChannelMention`
  free-chat path to build the in-memory session before passing
  to the runner) and `stripCaseUpdateBlock` /
  `markdownToMrkdwn` / `truncateForSlack` (case-file path).
- Gains `private readonly freeChatTurn: FreeChatTurnRunner` in
  the existing DI-aware ctor section.
- Two call sites updated: `this.runFreeChatTurn(args)` →
  `this.freeChatTurn.run(args)` (handleDmMessage + the channel-log
  free-chat closure inside handleChannelMention).

### Module map after this PR

| File | Lines | Purpose |
|---|---|---|
| `slack/index.ts` | **962** | adapter, event handlers, dispatcher, DM, channel mention |
| `slack/admin.ts` | 539 | `/pmk admin` (since v0.9) |
| **`slack/free-chat-turn.ts`** | **509** | **NEW: turn orchestration** |
| `slack/escalation.ts` | 308 | tranche 2 |
| `slack/presence.ts` | 146 | tranche 1 |
| `slack/context-retry.ts` | 132 | v0.11.1 |
| `slack/envelope-dedup.ts` | 45 | tranche 1 |
| `slack/concurrency.ts` | 36 | tranche 1 |
| `slack/progress.ts` | 23 | v0.10 |

### v0.13 progress

| Pre-v0.13 | After tranche 1 | After tranche 2 | After tranche 3 |
|---|---|---|---|
| 1708 | 1597 | 1411 | **962** |

-746 net (-44%) from baseline. Tranche 4 (dispatcher / event-
handler cleanup) will push under the 800-line coding-style max.
